# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MOSAP3 Procurement API",
        "description": "Documentação oficial da API do Sistema de Procurement MOSAP3.\n * \n *      Roles:\n *      - **Admin**: Acesso total ao sistema.\n *      - **Procurement Technician**: Gestão de cotações e fornecedores.\n *      - **Public (Fornecedor)**: Acesso via token único enviado por email.",
        "contact": {
            "email": "admin@mosap3.ao"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor da API"
        }
    ],
    "paths": {
        "/api/acquisitions": {
            "get": {
                "tags": [
                    "Aquisições"
                ],
                "summary": "Listar Aquisições",
                "description": "Lista todas as aquisições com filtros.",
                "operationId": "0ff44b8f44f4865443d39ae08f05170a",
                "parameters": [
                    {
                        "name": "supplier_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de aquisições"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/suppliers/{id}/acquisitions": {
            "get": {
                "tags": [
                    "Aquisições"
                ],
                "summary": "Histórico de Aquisições do Fornecedor",
                "operationId": "60134e48e42b17ce73901dac3534c228",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Histórico do fornecedor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/acquisitions/stats/products": {
            "get": {
                "tags": [
                    "Aquisições"
                ],
                "summary": "Produtos Mais Adquiridos",
                "description": "Retorna ranking de produtos mais comprados por quantidade ou valor total.",
                "operationId": "400fefbc74e2d90d21a553f0e3ec6b3f",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ranking de produtos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/audit-logs": {
            "get": {
                "tags": [
                    "Auditoria"
                ],
                "summary": "Listar logs de auditoria",
                "description": "Retorna a lista paginada de logs de auditoria com filtros opcionais.",
                "operationId": "9e62b5f5a6a0c15d4ae6dd3932622142",
                "parameters": [
                    {
                        "name": "event",
                        "in": "query",
                        "description": "Filtrar por evento",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filtrar por utilizador",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Data início (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Data fim (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Resultados por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de logs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/audit-logs/{id}": {
            "get": {
                "tags": [
                    "Auditoria"
                ],
                "summary": "Detalhe do log de auditoria",
                "operationId": "0d6d7c056e441ba591f9ac48024e1c65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Log encontrado"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}